requirements.txt
asgiref==3.9.2
black==25.9.0
click==8.3.0
colorama==0.4.6
Django==5.2.7
django-admin-cli==0.1.1
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
isort==6.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
